logging: add a nice little interface for doing logging that handles buffers
On the platform, there is an option to send serial messages in a blocking
fashion and non-blocking.  Neither of this are ideal as the blocking
will just make things not work (block forever) if you are not reading
data off the port.  The non-blocking results in consecutive calls to
logging resulting in lost data.  Here, we manage a buffer of log data
to be sent out.  We guarantee order but will drop messages if we can't
get the data sent or cannot send it quick enough.